+++ build.gradle
@@ -9,7 +9,7 @@ plugins {
 
 jar.archiveName = project.name + '.jar'
 // Add SNAPSHOT to make this publish as a beta.
-version '1.0.9'
+version '1.1.0'
 
 sourceCompatibility = 1.8
